AiDocumentHandler & AiSnippetHandler implemented
