+++ pebblo/app/service/loader/loader_doc_service.py
@@ -3,18 +3,15 @@ from pebblo.app.libs.responses import PebbloJsonResponse
 from pebblo.app.models.db_models import (
     AiDataModel,
     AiDataSource,
-    AiDocument,
-    AiSnippet,
     LoaderDocResponseModel,
     LoaderMetadata,
 )
 from pebblo.app.models.sqltables import (
     AiDataLoaderTable,
     AiDataSourceTable,
-    AiDocumentTable,
-    AiSnippetsTable,
 )
 from pebblo.app.service.discovery.common import get_or_create_app
+from pebblo.app.service.loader.document.document import AiDocumentHandler
 from pebblo.app.storage.sqlite_db import SQLiteClient
 from pebblo.app.utils.utils import get_current_time
 from pebblo.entity_classifier.entity_classifier import EntityClassifier
@@ -127,7 +124,8 @@ class AppLoaderDoc:
             logger.error(f"Get Classifier Response Failed, Exception: {e}")
             return doc_info
 
-    def _update_doc_details(self, doc, doc_info):
+    @staticmethod
+    def _update_doc_details(doc, doc_info):
         """
         Create a doc model and return its object
         """
@@ -165,183 +163,6 @@ class AppLoaderDoc:
         logger.info("Data Source Details has been updated successfully.")
         return data_source_obj.data
 
-    def _create_update_document_snippets(self, app_loader_details, data_source):
-        logger.info("Create update document snippet")
-        input_doc_list = self.data.get("docs", [])
-        existing_document = None
-        doc_obj = None
-        for doc in input_doc_list:
-            if not existing_document:
-                doc_obj = self._get_or_create_document(doc, data_source)
-                existing_document = doc_obj.data
-            snippet = self._create_snippet(doc, data_source, existing_document)
-            existing_document = self._update_document(existing_document, snippet)
-            app_loader_details = self._update_loader_documents(
-                app_loader_details, existing_document, snippet
-            )
-        return app_loader_details, existing_document, doc_obj
-
-    @staticmethod
-    def _count_entities_topics(restricted_data, doc_restricted_data, snippet_id):
-        logger.info("counting entities and topics started")
-        for data in doc_restricted_data:
-            # If entity in apps coll
-            if data in restricted_data:
-                # updating existing count and appending doc id
-                restricted_data[data]["count"] += doc_restricted_data.get(data, 0)
-                restricted_data.get(data, {}).get("docIds", []).append(snippet_id)
-            else:
-                # If entity or topic does not exist in  app coll
-                restricted_data[data] = {
-                    "count": doc_restricted_data.get(data, 0),
-                    "docIds": [snippet_id],
-                }
-                # Adding count and docId in app coll
-        logger.info("counting entities and topics finished.")
-        return restricted_data
-
-    def _update_loader_documents(self, app_loader_details, document, snippet):
-        logger.info("Updating Loader details with document and findings.")
-
-        # Updating documents value for AiDataLoader
-        documents = app_loader_details.get("documents", [])
-        documents.append(document.get("sourcePath"))
-
-        documents = list(set(documents))
-        app_loader_details["documents"] = documents
-
-        # Updating documentsWithFindings value for AiDataLoader
-        documents_with_findings = app_loader_details.get("documentsWithFindings", [])
-        if document.get("topics") not in ({}, None) or document.get("entities") not in (
-            {},
-            None,
-        ):
-            documents_with_findings.append(document.get("sourcePath"))
-            documents_with_findings = list(set(documents_with_findings))
-            app_loader_details["documentsWithFindings"] = documents_with_findings
-
-        # Updating source files in loaders
-        loader_info = app_loader_details.get("loaders", [])
-        if loader_info:
-            for loader in loader_info:
-                if loader.get("sourcePath") == document.get("loaderSourcePath"):
-                    if document.get("sourcePath") not in loader["sourceFiles"]:
-                        loader["sourceFiles"].append(document.get("sourcePath"))
-                loader["lastModified"] = get_current_time()
-
-        # Update doc entities & topics details from snippets
-        # Fetching entities and topics
-        entities_data = app_loader_details.get("docEntities", {})
-        topics_data = app_loader_details.get("docTopics", {})
-
-        if snippet.get("entities"):
-            # If entities exist in snippet
-            entities_data = self._count_entities_topics(
-                entities_data, snippet.get("entities"), snippet.get("id")
-            )
-        if snippet.get("topics"):
-            # If entities exist in snippet
-            topics_data = self._count_entities_topics(
-                topics_data, snippet.get("topics"), snippet.get("id")
-            )
-
-        app_loader_details["docEntities"] = entities_data
-        app_loader_details["docTopics"] = topics_data
-
-        # self.db.update_data(AiDataLoaderTable, app_loader_details)
-        logger.info("Loader details with document and findings updated successfully.")
-        return app_loader_details
-
-    def _update_document(self, document, snippet):
-        logger.info("Updating AIDocument with snippet reference")
-        existing_topics = document.get("topics")
-        if not existing_topics:
-            existing_topics = {}
-        existing_entities = document.get("entities")
-        if not existing_entities:
-            existing_entities = {}
-
-        topics = snippet.get("topics")
-        entities = snippet.get("entities")
-        if entities:
-            for entity in entities:
-                if entity in existing_entities.keys():
-                    updated_entity = existing_entities[entity]
-                    updated_entity["ref"].append(snippet.get("id"))
-                    existing_entities.update({entity: updated_entity})
-                else:
-                    existing_entities.update({entity: {"ref": [snippet.get("id")]}})
-        if topics:
-            for topic in topics:
-                if topic in existing_topics.keys():
-                    updated_topic = existing_topics[topic]
-                    updated_topic["ref"].append(snippet.get("id"))
-                    existing_topics.update({topic: updated_topic})
-                else:
-                    existing_topics.update({topic: {"ref": [snippet.get("id")]}})
-
-        document["topics"] = existing_topics
-        document["entities"] = existing_entities
-        logger.info("AIDocument Updated successfully with snippet reference")
-        return document
-
-    def _create_snippet(self, doc, data_source, document):
-        snippet_details = {
-            "appId": self.app_name,
-            "dataSourceId": data_source.get("id"),
-            "documentId": document.get("id"),
-            "metadata": {
-                "createdAt": get_current_time(),
-                "modifiedAt": get_current_time(),
-            },
-            "doc": doc.get("doc"),
-            # 'checksum': checksum,
-            "sourcePath": doc.get("source_path"),
-            "loaderSourcePath": data_source.get("sourcePath"),
-            "entities": doc.get("entities", {}),
-            "topics": doc.get("topics", {}),
-        }
-        ai_snippet_obj = AiSnippet(**snippet_details)
-        ai_snippet = ai_snippet_obj.dict()
-        status, snippet_obj = self.db.insert_data(AiSnippetsTable, ai_snippet)
-        logger.info("AISnippet created successfully.")
-        return snippet_obj.data
-
-    def _get_or_create_document(self, doc, data_source):
-        logger.info("Create or Update AIDocument")
-        filter_query = {
-            "appId": self.app_name,  # loadId or AppId ( Doubt)
-            "sourcePath": doc.get("source_path"),
-        }
-        status, output = self.db.query(AiDocumentTable, filter_query)
-        if output:
-            data = output.data
-            data["lastIngested"] = get_current_time()
-            data["metadata"]["updatedAt"] = get_current_time()
-            # self.db.update_data(AiDocumentTable, data)
-            return output
-        else:
-            metadata = {
-                "createdAt": get_current_time(),
-                "modifiedAt": get_current_time(),
-            }
-            # Document is not present, need to create.
-            ai_documents = {
-                "appId": self.app_name,
-                "dataSourceId": data_source.get("id"),
-                "metadata": metadata,
-                "sourcePath": doc.get("source_path"),
-                "loaderSourcePath": data_source.get("sourcePath"),
-                "owner": doc.get("file_owner"),
-                "userIdentities": doc.get("authorized_identities", []),
-                "lastIngested": get_current_time(),
-            }
-            ai_document_obj = AiDocument(**ai_documents)
-            ai_document_data = ai_document_obj.dict()
-
-            _, doc_obj = self.db.insert_data(AiDocumentTable, ai_document_data)
-            return doc_obj
-
     def process_request(self, data):
         try:
             self.db = SQLiteClient()
@@ -372,9 +193,9 @@ class AppLoaderDoc:
             data_source = self._create_data_source()
 
             # Iterate Each doc & Update AIDocument, AISnippets
-            app_loader_details, documents, doc_obj = (
-                self._create_update_document_snippets(app_loader_details, data_source)
-            )
+            document_handler = AiDocumentHandler(self.db, self.data)
+            app_loader_details, documents, doc_obj = document_handler.create_or_update_document(app_loader_details=app_loader_details,
+                                                                                                 data_source=data_source)
 
         except Exception as err:
             message = f"Loader Doc API Request failed, Error: {err}"
